+++ instrumentation/powerjob-4.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/powerjob/v_4_0/TestBasicFailProcessor.java
@@ -3,16 +3,16 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.powerjob.v_4_0_0;
+package io.opentelemetry.javaagent.instrumentation.powerjob.v_4_0;
 
 import tech.powerjob.worker.core.processor.ProcessResult;
 import tech.powerjob.worker.core.processor.TaskContext;
 import tech.powerjob.worker.core.processor.sdk.BasicProcessor;
 
-public class TestBasicProcessor implements BasicProcessor {
+public class TestBasicFailProcessor implements BasicProcessor {
 
   @Override
   public ProcessResult process(TaskContext context) {
-    return new ProcessResult(true, System.currentTimeMillis() + "success");
+    return new ProcessResult(false, "fail");
   }
 }
